finetuning/pyproject.toml
[tool.poetry]
name = "finetuning"
version = "0.1.0"
description = ""
authors = ["dewasahu2003 <dewasahu2003@gmail.com>"]
readme = "README.md"
packages = [{include = "finetuning"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
numpy = "1.24.0"
pandas = "2.2.2"
peft = "0.11.0"
datasets = "2.19.1"
transformers = "4.40.2"
safetensors = "0.4.3"
comet-ml = "3.42.0"
accelerate = "0.30.1"
bitsandbytes = "0.42.0"
pydantic-settings = "2.2.1"
scikit-learn = "1.4.2"
qwak-sdk = "0.5.82"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
